requirements.txt
alembic==1.9.4
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.5
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
python-dotenv~=1.0.0